+++ packages/components/atoms/f-form-field/CHANGELOG.md
@@ -4,6 +4,25 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v1.16.1
+------------------------------
+*June 30, 2021*
+
+### Changed
+- Label description class name.
+- Label description from `<p>` to `<span>`.
+- Background colours to `common.scss`
+
+
+v1.16.0
+------------------------------
+*June 28, 2021*
+
+### Fixed
+- Hover state to be consistent across all field types.
+- First grouped form field to have correct `border-radius`.
+
+
 v1.15.1
 ------------------------------
 *June 15, 2021*